test(formatter/sort_imports): Split test files (#17977)
Pure refactoring; Just split 1 test file -> 2 files by category.

(Will merge without review)
