+++ crates/oxc_formatter/tests/ir_transform/sort_imports/mod.rs
@@ -1,0 +1,2 @@
+mod basic;
+mod groups;
